refactor(web): extract useScheduleData composable
